Refactor changes according to comments
