+++ src/config/constants.ts
@@ -133,6 +133,7 @@ export const PERCENTAGE_100 = new BigNumber(PERCENTAGE);
 export const DEFAULT_BAKER_FOR_NON_TEZ_LP = null;
 export const DEFAULT_STABLESWAP_POOL_ID = 0;
 
+export const STABLESWAP_FARM_ACCUM_PRECISION = 10_000_000_000;
 //#endregion
 
 // error constant
