+++ src/modules/farming/interfaces/index.ts
@@ -1,5 +1,3 @@
-export * from '@shared/types';
-
 export * from './farming-contract.interface';
 export * from './farming-storage.interfaces';
 export * from './farming.interfaces';
