+++ src/modules/stableswap/types/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
 export * from './mapped-api';
 export * from './stableswap';
+export * from './stableswap.contract.types';
